Tweaks to object-creation methods.
git-svn-id: svn+ssh://svn.internal.sanger.ac.uk/repos/svn/pathsoft/genedb-ng/trunk@8670 ee4ac58c-ac51-4696-9907-e4b3aa274f04

+++ genedb-db/src/org/gmod/schema/mapped/FeatureLoc.java
@@ -95,12 +95,12 @@ public class FeatureLoc implements Serializable {
         this.rank = rank;
     }
 
-    public FeatureLoc(Feature sourceFeature, Feature feature, int fmin, int fmax, short strand, int phase) {
-        this(sourceFeature, feature, fmin, false, fmax, false, strand, phase, 0, 0);
+    public FeatureLoc(Feature sourceFeature, Feature feature, int fmin, int fmax, int strand, int phase) {
+        this(sourceFeature, feature, fmin, false, fmax, false, (short) strand, phase, 0, 0);
     }
 
-    public FeatureLoc(Feature sourceFeature, Feature feature, int fmin, int fmax, short strand, int phase, int rank) {
-        this(sourceFeature, feature, fmin, false, fmax, false, strand, phase, 0, rank);
+    public FeatureLoc(Feature sourceFeature, Feature feature, int fmin, int fmax, int strand, int phase, int rank) {
+        this(sourceFeature, feature, fmin, false, fmax, false, (short) strand, phase, 0, rank);
     }
 
     FeatureLoc(Feature parent, Feature child, StrandedLocation location) {
